[docker] Copy missing libraries into PySide6's folder
`libQt63DQuickScene3D.so` is missing from the default PySide6 install
despite being present in the Qt 6.8.3 install.

If it is missing, the "Application" part of Meshroom cannot be displayed
as a "Component not ready" error is thrown. The missing library is thus
copied from the Qt install folder, and its own dependencies are copied
with it to prevent issues when looking for them.